Fonction modification / suppression

+++ BLL/GestionProduits.cs
@@ -17,6 +17,16 @@ namespace BLL
             return uneGestionProduits;
         }
 
+        public static Categorie GetCategorieById(int id)
+        {
+            return DAL.CategorieDAO.GetCategorieById(id);
+        }
+
+        public static Categorie GetCategorieByNom(string nom)
+        {
+            return DAL.CategorieDAO.GetCategorieByNom(nom);
+        }
+
         public static void SetchaineConnexion(string chaine)
         {
             ConnexionBD.GetConnexionBD().SetchaineConnexion(chaine);
@@ -35,7 +45,12 @@ namespace BLL
 
         public static int SupprimerProduit(int code)
         {
+            if (ProduitDAO.ProduitEstUtilise(code) == false) { 
             return ProduitDAO.DeleteProduit(code);
+            }
+            else             {
+                throw new Exception("Le produit est utilisé et ne peut pas être supprimé.");
+            }
         }
     }
 }
